split code_803D0D8
