+++ ld_script.txt
@@ -197,6 +197,7 @@ SECTIONS {
         src/personality_test2.o(.text);
         src/code_803D0D8.o(.text);
         src/code_803D110.o(.text);
+        src/code_803DA7C.o(.text);
         src/code_803E46C.o(.text);
         src/code_803E724.o(.text);
         src/code_8040094.o(.text);
@@ -557,6 +558,7 @@ SECTIONS {
         src/dungeon_config.o(.rodata);
         data/data_80F59C8_1.o(.rodata);
         src/code_803D110.o(.rodata);
+        src/code_803DA7C.o(.rodata);
         data/data_80F59C8_3.o(.rodata);
         src/code_803E724.o(.rodata);
         data/data_80F59C8_0.o(.rodata);
